+++ NEWS.md
@@ -1,5 +1,6 @@
 # pkgdown (development version)
 
+* Anywhere you can use `_pkgdown.yml`, you can now use `_pkgdown.yaml` (#2244).
 * `build_article()` no longer generates the wrong source link when you build your site outside of the root directory (#2172).
 * `build_reference()` matches usage for S3 and S4 methods to the style used by R 4.0.0 and later (#2187).
 * `<source>` tags now have their `srcref` attributes tweaked in the same way that the `src` attributes of `<img>` tags are (#2402).